+++ backend/src/main/java/com/example/smu_club/club/service/BatchClubService.java
@@ -1,0 +1,29 @@
+package com.example.smu_club.club.service;
+
+import com.example.smu_club.club.repository.ClubMemberRepository;
+import com.example.smu_club.domain.ClubMember;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class BatchClubService {
+    private final ClubMemberRepository clubMemberRepository;
+
+    @Transactional
+    public List<ClubMember> findExpiredClubMembers() {
+        //1. 모집 종료된지 1달이 지난 동아리 회원 조회
+        // 오늘 날짜에서 1달 이전 날짜 >= 동아리 모집 종료일 이라면, 삭제 대상으로 간주한다.
+        return clubMemberRepository.findExpiredClubMembers(LocalDateTime.now().minusMonths(1));
+    }
+
+
+    public int cleanupExpiredClubMembers(List<ClubMember> expiredClubMembers) {
+        //2. 만료된 동아리 회원 삭제 (배치 삭제)
+        return clubMemberRepository.deleteAllInBatchWithCount(expiredClubMembers);
+    }
+}
